# `UnmarshalJSON`, `Scan` and `NewFromString` performance improvements